HEY-17: feat(question): ensure actions archive a question only a person that created

+++ app/Http/Controllers/QuestionController.php
@@ -93,8 +93,13 @@ class QuestionController extends Controller
         return back();
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function archive(Question $question): RedirectResponse
     {
+        $this->authorize('archive', $question);
+
         $question->delete();
 
         return back();
